backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
drf-spectacular>=0.27
pydantic>=2.7
python-dotenv>=1.0
requests>=2.32
psycopg2-binary>=2.9
